--- src/Mafia.hs
@@ -1,2 +1,0 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-module Mafia where
